Add a stiction coefficient.
Also adjust the graph to show the force.

These numbers are not looking sensible, particularly at startup from
stopped (which is important here). The force goes way too high (I don't
believe T=E/ω remains true at low speeds) which means the stiction
effects are unobservable. There should be a point at which the PWM is
nonzero but the vehicle is not moving because of the stiction, but I'm
not seeing that.